perbaikan fitur hapus file

+++ app/Http/Controllers/StudentController.php
@@ -2,6 +2,7 @@
  
 namespace App\Http\Controllers;
  
+use Illuminate\support\Facades\File;
 use Illuminate\Http\Request;
 use App\Models\Student; //add Student Model - Data is coming from the database via Model.
  
@@ -24,6 +25,9 @@ class StudentController extends Controller
 
     public function store(Request $request)
     {
+        $request->validate([
+            'gender' => ['in:L,P'],
+        ]);
         $newName = '';
 
         if($request->file('photo')){
@@ -54,7 +58,9 @@ class StudentController extends Controller
     {
         $student = Student::find($id);
         $newName = '';
-        if($request->file('image')->isValid()){
+
+        if($request->image){
+            $oldImage = $student->image;
             $extension = $request->file('image')->getClientOriginalExtension();
             $newName = $request->name.'-'.now()->timestamp.'.'.$extension;
             $request->file('image')->storeAs('gambar',$newName);
@@ -64,17 +70,30 @@ class StudentController extends Controller
             // 2. update datanya
 
             // kalau gagal upload, langsung return error
+            $student->update($request->except('image') + ['image' => $newName]);
+           
+            file::delete(public_path().'/storage/gambar/'.$oldImage);
+            }
+            return redirect('students')->with('flash_message', 'student Updated!');  
         }
 
-        $student->update($request->except('image') + ['image' => $newName]);
-
-        return redirect('students')->with('flash_message', 'student Updated!');  
-    }
+       
+        // if ($upload) {
+        //     $delete = File::delete(public_path() . '/storage/photo/' . $oldImage);
+        //     $student->update($request->except('image') + ['image' => $newName]);
+        
+    
  
 
     public function destroy($id)
     {
-        Student::destroy($id);
+        $student = Student::find($id);
+        $oldImage = $student->image;
+        $delete = file::delete(public_path() . '/storage/gambar/' . $oldImage);
+        if ($delete) {
+            Student::destroy($id);
+        }
+        
         return redirect('students')->with('flash_message', 'Student deleted!');  
     }
 }
